test(sass): matchers refactoring for GMock-like
