+++ tests/python/test/sass/CMakeLists.txt
@@ -3,4 +3,5 @@ add_pytest(atomic)
 add_pytest(branch)
 add_pytest(load)
 add_pytest(matchers)
+add_pytest(matchers_impl)
 add_pytest(sass)
